Trying to sync up with Jeremy Way app

+++ app/Http/Controllers/JobController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Models\Job;
 use App\Models\User;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 
@@ -12,7 +11,7 @@ class JobController extends Controller
 {
     public function index()
     {
-        $jobs = Job::with('employer')->latest()->paginate(3);
+        $jobs = Job::with('employer')->latest()->simplePaginate(3);
 
         return view('jobs.index', [
             'jobs' => $jobs
@@ -39,23 +38,20 @@ class JobController extends Controller
         Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
-            'employer_id' => 1,
+            'employer_id' => 1
         ]);
 
-        return redirect('jobs');
+        return redirect('/jobs');
     }
 
     public function edit(Job $job)
     {
-        Gate::authorize('edit-job', $job);
-
         return view('jobs.edit', ['job' => $job]);
     }
 
     public function update(Job $job)
     {
-
-        // authorize (on hold...)
+        Gate::authorize('edit-job', $job);
 
         request()->validate([
             'title' => ['required', 'min:3'],
@@ -72,7 +68,7 @@ class JobController extends Controller
 
     public function destroy(Job $job)
     {
-        // authorize(on hold...)
+        Gate::authorize('edit-job', $job);
 
         $job->delete();
